+++ CHANGELOG.md
@@ -172,3 +172,6 @@
 ### (2015/5/8)
 - Add setInterval ajax event to get summary instead of socket.io
 - Add TreeWeb to generate shortest paths to go to activities
+
+### (2015/5/9)
+- Send paths to nata-server